\name{calculateCt}
\alias{calculateCt}
\title{Calculate Ct Values}
\description{Calculate Ct values from Delta Rn matrices.}
\usage{
calculateCt(x, threshold, start.cycle)
}
\arguments{
\item{x}{a Delta Rn matrix as generated by \code{\link{calculateDeltaRn}}.}
\item{threshold}{a threshold cycle value.}
\item{start.cycle}{the minimum cycle to use when determining Ct.}
}
\value{A vector of Ct values.}
\seealso{
\code{\link{calculateDeltaRn}} for generating Delta Rn matrices and \code{\link{calculateThreshold}} for calculating threshold values.
}
