\name{calculateDeltaRn}
\alias{calculateDeltaRn}
\title{Calculate Delta Rn Values}
\description{Calculate Delta Rn values from an Rn matrix and a baseline data.frame.}
\usage{
calculateDeltaRn(x, b)
}
\arguments{
\item{x}{an Rn matrix as generated by \code{\link{calculateRn}}.}
\item{b}{a baseline data.frame as generated by \code{\link{calculateBaseline}}.}
}
\value{A matrix of Delta Rn values.}
\seealso{
\code{\link{calculateRn}} for generating Rn matrices and \code{\link{calculateBaseline}} for calculating baseline values.
}
