\name{calculateRn}
\alias{calculateRn}
\title{Calculate Rn Values from a dye pair object}
\description{Calculate the Rn values for all wells in a dyepair object.}
\usage{
calculateRn(x)
}
\arguments{
\item{x}{a dyepair list as created by \code{\link{fluorescence}}.}
}
\value{A matrix of Rn values.}
\seealso{
\code{\link{fluorescence}}.
}
