\name{calculateThreshold}
\alias{calculateThreshold}
\title{Calculate Threshold Values from Baseline Data}
\description{Calculate a card threshold value based on the average baseline standard deviation.}
\usage{
calculateThreshold(x)
}
\arguments{
\item{x}{a baseline data.frame as created by \code{\link{calculateBaseline}}.}
}
\value{The calculated threshold value.}
\details{
A single threshold value is generated by taking the mean of \eqn{baseline+(10*sd)} across all wells.
}
\seealso{
\code{\link{calculateBaseline}} for generating baseline data from Rn matrices.
}
