\name{pcrModelFlatPhaseEnd}
\alias{pcrModelFlatPhaseEnd}
\title{Determine Flat Phase End Points}
\description{Determine flat phase endpoints for each model in a list generated by \code{\link{pcrModel}}.}
\usage{
pcrModelFlatPhaseEnd(x, min=5, offset=2, default=15, mc.cores=1, ...)
}
\arguments{
\item{x}{a model list as created by \code{\link{pcrModel}}.}
\item{min}{the minimum cycle to report as an end.}
\item{offset}{the cycle offset to use.}
\item{default}{the default cycle to report when takeoff calculation fails.}
\item{mc.cores}{the number of cores used for processing.}
\item{...}{further arguments to pass to \code{\link{takeoff}}}
}
\value{A vector of calculated plat phase end cycles.}
\seealso{
\code{\link{pcrModel}} for generating model lists from fluorescence data.
\code{\link{pcrModelTakeoff}} for the underlying function.
}
