\name{pcrModelTakeoff}
\alias{pcrModelTakeoff}
\title{Determine Takeoff Points from Models}
\description{Extract PCR takeoff points for each model in a list generated by \code{\link{pcrModel}}.}
\usage{
pcrModelTakeoff(x, mc.cores=1, ...)
}
\arguments{
\item{x}{a model list as created by \code{\link{pcrModel}}.}
\item{mc.cores}{the number of cores used for processing.}
\item{...}{further arguments to pass to \code{\link{takeoff}}}
}
\value{A vector of calculated takeoff cycles.}
\seealso{
\code{\link{pcrModel}} for generating model lists from fluorescence data.
\code{\link{takeoff}} for the underlying function.
}
