\name{plotFluorescence}
\alias{plotFluorescence}
\title{Plot Fluorescence data}
\description{Generate plots of Fluorescence values.}
\usage{
plotFluorescence(x, wells=NA, col=dye.cols, ylim=NA, ...)
}
\arguments{
\item{x}{a dyepair matrix as generated by \code{\link{fluorescence}}.}
\item{wells}{the wells to plot.}
\item{col}{the dye colours to use.}
\item{ylim}{the y limits to use.}
\item{...}{further arguments to pass to \code{\link{lines}}}
}
\seealso{
\code{\link{fluorescence}} for generating fluorescence data.
}
