\name{plotRn}
\alias{plotRn}
\title{Plot Rn data}
\description{Generate plots of Rn or Delta Rn matrices.}
\usage{
plotRn(x, wells=NA, col='blue', ylim=NA, log=FALSE, ...)
}
\arguments{
\item{x}{an Rn matrix as generated by \code{\link{calculateRn}} or \code{\link{calculateDeltaRn}}.}
\item{wells}{the wells to plot.}
\item{col}{the colour used for the Rn curves.}
\item{ylim}{the y limits to use.}
\item{log}{logical. Should the Rn data be logged before plotting?.}
\item{...}{further arguments to pass to \code{\link{lines}}}
}
\seealso{
\code{\link{calculateRn}} and \code{\link{calculateDeltaRn}} for generating Rn and dRn matrices.
}
