\name{readMulticomponent}
\alias{readMulticomponent}
\title{Read multicomponent files}
\description{Read multicomponent files for analysis.}
\usage{
readMulticomponent(file, sample)
}
\arguments{
\item{file}{input filename to read.}
\item{sample}{sample identifier.}
}
\value{A data.frame of multicomponent data.}
